requirements.txt
psycopg2-binary==2.9.5
Django==4.1.4
django-filter==22.1
djangorestframework==3.14.0
drf-yasg==1.21.4
